Enemies and bullets are now destroyed on collision
