+++ src/main.rs
@@ -6,6 +6,6 @@ use bevy::prelude::*;
 fn main() {
     App::new()
         .add_plugins(DefaultPlugins)
-        .add_plugin(game::Game)
+        .add_plugin(game::game::Game)
         .run();
 }
